gdb/aarch64: restore in-order watchpoint matching
At Red Hat we have an out of tree AArch64 watchpoint test which broke
after this commit:

  commit cf16ab724a41e4cbaf723b5633d4e7b29f61372b
  Date:   Tue Mar 12 17:08:18 2024 +0100

      [gdb/tdep] Fix gdb.base/watch-bitfields.exp on aarch64

The problem with AArch64 hardware watchpoints is that they (as I
understand it) are restricted to a minimum of 8 bytes.  This means
that, if the thing you are watching is less than 8-bytes, then there
is always scope for invalid watchpoint triggers caused by activity in
the part of the 8-bytes that are not being watched.

Or, as is the case in this RH test, multiple watchpoint are created
within an 8-byte region, and GDB can miss-identify which watchpoint
actually triggered.

Prior to the above commit the RH test was passing, however, the test
was relying on, in the case of ambiguity, GDB selecting the first
created watchpoint.  That behaviour changed with the above commit.
Now GDB favours reporting non write breakpoints, and will only report
a write breakpoint if not non-write breakpoint exists in the same
region.

I fully understand that this test could be rewritten so that the
_correct_ behaviour requires GDB to report the later watchpoint, so to
some degree, the change I propose here is pretty arbitrary....

... but I'd like to restore the old behaviour (that reports
watchpoints in creation order), while also retaining the improvements
added in the above commit.

The key change from the above commit is that
aarch64_stopped_data_address should return false if multiple write
watchpoints match, and there are no read watchpoints at the same
location.  This tells GDB to test all write watchpoints and figure out
which one changed.

So, the new strategy is:

  - Remove the two phase loop, everything is now handled in one pass.

  - If we don't care about _which_ watchpoint matched, we just want to
    know if _any_ watchpoint matched, then return true as soon as we
    find a match, this is unchanged from the above commit.

  - Record the address of the first matching watchpoint.

  - Count the number of matching read and write watchpoints.

  - After checking all watchpoints, if we have no matches, or multiple
    write matches and no read watchpoints, then return false, GDB will
    then manually check all write watchpoints to see which changed.

  - Otherwise, we have at least one match, this could be a single
    write watchpoint, or any number of write watchpoints and at least
    one read watchpoint.  In this case, return true along with the
    address of the first matching watchpoint, this restores the
    behaviour from before the above commit.