[wip] gdb: don't try to add unreadable [vsyscall] page to core dump
Currently GDB will always try to dump the [vdso] and [vsyscall]
mappings, which mirrors the kernel's behaviour.  However, with more
recent kernels the [vsyscall] mapping is read-only, which prevents GDB
from adding its contents to the core dump.  This also causes GDB to
emit a warning like this:

  warning: Memory read failed for corefile section, 4096 bytes at 0xffffffffff600000.

Lets prevent the warning by not trying to dump the [vdso] and/or
[vsyscall] if the mapping is not readable.

This solution isn't ideal.  A kernel core dump does include these
mappings, along with the contents of the mapping, so ideally, GDB
would include this mapping too.  Maybe there's an alternative approach
we can find that would allow us to read the mapping?

Or maybe we should take this back to the kernel.  If the mapping is
included in the core dump then making it executable, but not readable,
is really not protecting anything, but it is impacting the debugger.