[wip] gdb: don't try to add unreadable [vsyscall] page to core dump
============================================================

NOTE: I don't think this patch should be going upstream.  Currently
the new warning that is added will be printed twice because we're
printing it within the function that is really about iterating over
the mappings, while the warning really relates to what we need to _do_
with the mapping.

The problem is that the warning comes from the
gdbarch_find_memory_regions call, but that is called multiple times.
We really only want to emit the warning when we realise that the
section cannot be read, and when we know that we do plan to read the
section, which I think we might be able to figure out in
gcore_create_callback.

Maybe if we passed an (optional?) mapping name to
gcore_create_callback, then we could use that new name to give a
better warning from within gcore_create_callback?

============================================================

Currently GDB will always try to dump the [vdso] and [vsyscall]
mappings, which mirrors the kernel's behaviour.  However, with more
recent kernels, on x86-64 at least, the [vsyscall] mapping is
excutable, but not readable, which prevents GDB from adding its
contents to the core dump.  This also causes GDB to emit a warning
like this:

  warning: Memory read failed for corefile section, 4096 bytes at 0xffffffffff600000.

Lets prevent the warning by not trying to dump the [vdso] and/or
[vsyscall] if the mapping is not readable.  What I'm actually
proposing is that we give a more detailed, case specific warning,
like:

  warning: unable to dump unreadable [vdso] mapping

This solution isn't ideal.  A kernel core dump does include these
mappings, along with the contents of the mapping, so ideally, GDB
would include this mapping too.  Maybe there's an alternative approach
we can find that would allow us to read the mapping?  I suspect not
though.  Surely the reason the mapping is unreadable is to prevent the
contents being read, and if GDB could read the contents, then others
could read the contents, and this would render the whole unreadable
mapping pointless.