gdb: reduce generated core file size
Bug PR gdb/33222 points out that the core files generated by GDB can
be significantly larger than those generated by the kernel.  The bug
goes on to identify guard pages, those mapped as PROT_NONE (neither
readable, writable, or executable) as being a problem, GDB includes
the contents of these pages (all zeros) within the generated core
files, while the kernel does not.

Using the test from this patch, here's a program header line, (read
using readelf), from a kernel generated core file, for a PROT_NONE
mapping:

  Type           Offset   VirtAddr           PhysAddr           FileSiz  MemSiz   Flg Align
  LOAD           0x039000 0x00007f2bdc0b3000 0x0000000000000000 0x000000 0x001000     0x1000

And here's the same mapping for a core file generated by GDB:

  Type           Offset   VirtAddr           PhysAddr           FileSiz  MemSiz   Flg Align
  LOAD           0x05c430 0x00007f2bdc0b3000 0x0000000000000000 0x001000 0x001000 R   0x1

The difference between these lines that really matters is the FileSiz
field.  From the kernel core file this is 0, while for the GDB
generated core this is 0x1000.

The kernel's zero size makes sense.  This PROT_NONE mapping was never
accessed by the inferior, so no memory was ever allocated by the
kernel for this mapping.  As such, there's really no reason to include
any content in the core file.

Addressing this problem requires changes in two area:

1. Fixing gcore_create_callback to handle PROT_NONE mappings:
=============================================================

In gcore_create_callback, we already have an `if` block that is
guarded like this:

   if (!read && !write && !exec && !modified)
     {
       ...
     }

This block should trigger for any mapping that is PROT_NONE, and
unmodified, which you'd think our guard pages should be (more on this
later).  Within this `if` block, we immediately return, which will
prevent the mapping ever being created.  This isn't what we want, as
this doesn't match the kernels behaviour, where the mapping is created
in the core file with no contents.

To fix this, within the above `if` block, instead of just returning, I
instead update the section flags to remove the SEC_LOAD and
SEC_HAS_CONTENTS flags.  The result of this change is that any mapping
that triggers the above `if` block will result in a program header
entry with a zero FileSiz, and this will reduce the core file size.

You might wonder how I can be so confident making this change to the
above `if` block.  Isn't this going to break some existing behaviour
in GDB?  I also mentioned about that you'd think our guard pages
should already trigger this block, right?

The answer to both questions is that the above `if` block is never
triggered!

The MODIFIED flag, it turns out, is always true.  Every code path that
can call gcore_create_callback, always passes true for the MODIFIED
argument.  Which leads us nicely to the next part of the fix:

2. Identify Mappings That Are Unmodified
========================================

What we need then, is for the Linux arch/target code to pass false for
those mappings that are unmodified.

But how can we identify such mappings?

I think the answer lies in better parsing of the /proc/PID/smaps file,
which we already parse within linux-tdep.c as part of the core file
creation process, and making better use of the information we already
have to hand.  Specifically, I think two pieces of information can
help us.

First, there is the Rss field.  We don't currently parse this, but we
could.  As I understand it, this field represents the physical memory
that has been allocated for a mapping.  Assuming a 4k page size, then
a 1 page anonymous mapping will initially have a Size of 4k, but an
Rss of 0; no physical memory has yet been allocated.

When the inferior reads or writes into the mapping then pages are
actually allocated, and the Rss value will increase to 4k.

In the case of guard pages that are allocated PROT_NONE, these pages
can never be accessed, and so their Rss will remain 0 for their entire
lifetime.

As such, a mapping with an Rss of 0 can be considered unmodified.

The Anonymous field indicates physical memory that has been allocated,
but which is not associated with any file.  Taking our example from
above, the 4k anonymous mapping which has never been access has a Size
of 4k, an Rss of 0, and an Anonymous value also of 0.

But more interestingly, a private, file backed mapping, will also have
an Anonymous value of 0 so long as the file backed mapping has not
been written too.  This is useful because a read-only mapping from a
file need not be included in the core file.

What I propose is a "fail safe" strategy.  We assume that all mappings
are modified by default, which retains the existing behaviour.  But,
if, when we parse the smaps file, we see an Rss or Anonymous value of
zero, then we can record that the mapping is unmodified and this can
then be passed to the gcore_create_callback function, which will allow
the `if` block, discussed in (1) above, to trigger.

On (much) older kernels, the smaps file didn't exist, and we fall back
to using the /proc/PID/maps file instead.  If this happens then,
because of our fail safe strategy, mappings will be emitted just as
they always have.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=33222