[wip] gdb: attempt to reduce generated core file size
See the comments in this commit, but the 'modified' argument is always
set to true.  Can we read a better value from the smaps file?

Can we safely assume that the modified flag should be ignored?

If we ignore it, then we can skip adding content for some segments,
this is a huge improvement.

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=33222