WIP: remove dead code path
As target_preopen unpushes any existing core_target, then we cannot
possibly have a core file loading into the current program space (or
at least we shouldn't).

This could be extended with an extra assert:

  gdb_assert (current_program_space->core_bfd () == nullptr);

But that assert would then be removed in the next commit, so adding
that might be a little pointless?