[wip] gdb: have gdbarch create the core_target ... or a sub-class
Instead of just creating a core_target object, have the gdbarch create
the core_target for us.

This opens the door to gdbarch creating a sub-class of core_target,
though doing this will require the core_target class moving into a
header file, that's not done here, but can be done in a later commit.

The benefit of this is that core_target sub-classes can implement
methods which would override new target methods that can then be
added.

For example, linux_core_file_address_in_memtag_page, which is
currently a free function, could become a target method.  Then
architectures that support memtags can override this in a core target
sub-class add functionality in this way.