gdb: remove program_space::core_bfd member function
This commit cleans up the hack that was left in as part of the last
commit and removes the  program_space::core_bfd member function.

In every place that 'current_program_space->core_bfd ()' was used I
now call 'get_inferior_core_bfd (current_inferior ())'.

I think there is further scope for improving things in the future, we
could move some of the functionality that currently accesses the
global core file state into new member function for the core_target
class.  But that's not done in this commit.

But I think in some places, at the top level (e.g. user command
functions), there's always going to be a few cases where the cleanest
solution will be to just grab the core file BFD directly from the
current inferior, so some of these changes will be long term.

There should be no user visible changes after this commit.