gdb: add core file name to 'info inferiors' output
This commit builds on the previous commit.  In the future I am
proposing to move the core file BFD from the program_space into the
core_target.  In the last commit I updated 'maint info program-spaces'
to remove the core file name from the output.

In this commit I'm adding the core file name to the 'info inferiors'
output.

My proposal right now is to add the 'Core File' as a new (fifth)
column to the 'info inferiors' output.

In order to save space, this column is only added when one of the
current inferiors has a core file loaded.  If there are no core files
loaded then the 'Core File' column is dropped.

The problem with this approach is that the 'info inferiors' output now
has a column for both 'Executable' and 'Core File', both of which can
be pretty long.  For example, run the test
'gdb.multi/multi-target-info-inferiors.exp', and look at examples of
'info inferiors' in the gdb.log.  In a standard 80 column terminal
it's not uncommon for the 'Core File' column to end up being line
wrapped, which really isn't great.  Here's an example in an 80 column
terminal, with some long file paths, wrapped exactly as I see it:

    Num  Description       Connection           Executable
                                                                Core File
  * 1    process 3322840   1 (core)             /home/andrew/projects/binutils-gdb
  /build/gdb/testsuite/outputs/gdb.base/corefile/corefile /home/andrew/projects/bi
  nutils-gdb/build/gdb/testsuite/outputs/gdb.base/corefile/corefile.core

Where there's only one long filename things somehow don't seem so bad,
here's the old 'info inferiors' output:

    Num  Description       Connection           Executable
  * 1    process 3322840   1 (core)             /home/andrew/projects/binutils-gdb
  /build/gdb/testsuite/outputs/gdb.base/corefile/corefile

I think because the column names don't get wrapped it somehow feels
better.

One alternative would be to just add a whole new command, 'info
core-files', with output like:

    Num  Core File
  * 1    /home/andrew/projects/binutils-gdb/build/gdb/testsuite/outputs/gdb.base/c
  orefile/corefile.core

The 'Num' column here would just be the inferior number again.  In
effect this is new command is just an addendum to 'info inferiors'.

Another alternative would be to add the core file beneath the inferior
line in the 'info inferiors' output, like this:

    Num  Description       Connection           Executable
  * 1    process 3322840   1 (core)             /home/andrew/projects/binutils-gdb
  /build/gdb/testsuite/outputs/gdb.base/corefile/corefile
    Core File: /home/andrew/projects/binutils-gdb/build/gdb/testsuite/outputs/gdb.
  base/corefile/corefile.core

We already support some trailing information like this related to
vfork parent/child relationships, so maybe adding the core file here
would be OK?

I'd love to hear what people think about the ideas above, or if there
are any other ideas that people have for where to display this
information.

I extended gdb.base/corefile.exp to check the current output style,
and updated the gdb.multi/multi-target-info-inferiors.exp test to take
the new output into account.