gdb: fix permissions for unreadable mappings generated by gcore
This commit follows the previous, and makes a small fix to the program
header permissions within a GDB generated core file.

The previous commit added a test which creates PROT_NONE memory
mappings, that is, mappings that are neither readable, writable, or
executable.  Using readelf to examine the program headers, in a kernel
generated core file, for the PROT_NONE mapping I see:

  Type           Offset   VirtAddr           PhysAddr           FileSiz  MemSiz   Flg Align
  LOAD           0x039000 0x00007f2bdc0b3000 0x0000000000000000 0x000000 0x001000     0x1000

While for a GDB generated core file, after the last commit, I see:

  Type           Offset   VirtAddr           PhysAddr           FileSiz  MemSiz   Flg Align
  LOAD           0x05c430 0x00007f2bdc0b3000 0x0000000000000000 0x000000 0x001000 R   0x1

My interest in this commit is the Flg field.  The kernel generated
core file correctly shows this mapping with no flags, this indicates
the mapping is neither readable, writable, or executable.

Meanwhile, GDB incorrectly shows the mapping as readable.

Now this is all rather academic at this point.  GDB is able to read
even these unreadable mappings, both in a live inferior, and in both
core files.  So fixing this permissions flag might seem pointless.

However, I think there are two possible benefits to fixing this.
First, we don't know if there are other tools that examine core files
that might rely on accurate permission flags, and second, though I
don't think these permissions are user visible in GDB right now, I
would like to make them user visible in the future, by having 'info
proc mappings' for a core file target be closer to 'info proc
mappings' for a live target, e.g. include the Perms (permissions)
column.

Fixing this isn't trivial, but isn't that hard.  The problem is that
BFD doesn't have a concept of a section being unreadable.  The most
related section flag is SEC_READONLY, which removes the writable
permission, but the readable permission is assumed to always be
needed.

To work around this I propose making use of BFD's section userdata.

During core file creation GDB parses the inferior's mappings, and
creates a section to represent each one.  Later, each section is used
to create a program header.

If at the point the section is created we can see that the mapping is
PROT_NONE, then we add some user data to the section, this userdata
will indicate that the section is unreadable.

When we later create the program header, if a section has userdata,
and the userdata indicates an unreadable section, then the program
header will be created without the readable flag.

There is just one small problem with this approach.  The BFD object
used to create the core file is managed via GDB's gdb_bfd.{h,c}
mechanism, see gcore_command and the use of gdb_bfd_ref_ptr.  What
this means is that when the BFD is closed, it is done so via
gdb_bfd_unref in gdb_bfd.c.

Now gdb_bfd.c already makes use of the section's userdata field.  But
this is only for when we read in a BFD as part of the DWARF parser.
However, when closing a BFD in gdb_bfd.c, if a section has any
userdata, then we assume it is of type 'struct gdb_bfd_section_data',
and try to deallocate it as such.

With the changes I propose in this patch, adding a new type of
userdata, this results in a problem.  When the core file BFD is closed
we end up trying to deallocate my new 'struct gcore_section_userdata'
as if it were a 'struct gdb_bfd_section_data' which is clearly wrong,
and results in a crash.

The easiest solution I could see to this problem was just to ensure
that the userdata never escaped the core file creation process.  So I
extended the SCOPE_EXIT block that already existing in
write_gcore_file, and had this discard any per-section userdata.

With this done I updated the existing test to check for the expected
permissions.