gdb: notify of inferior switch when needed from 'thread' command
While working on the commit:

  commit 9959019545d8d6d71d927f20f088efba944b1e9c
  Date:   Sun Sep 28 16:16:53 2025 +0100

      gdb: fix for 'set suppress-cli-notifications on' missed case

I spotted this message in the gdb.mi/user-selected-context-sync.exp
test script:

  # Idea for the future: selecting a thread in a different inferior.  For now,
  # GDB doesn't show an inferior switch, but if it did, it would be a nice
  # place to test it.

What this message is talking about is this behaviour:

  (gdb) info threads
    Id   Target Id                              Frame
    1.1  Thread 0xf7dbc700 (LWP 818430) "thr" 0xf7eb2888 in clone () from /lib/libc.so.6
    1.2  Thread 0xf7dbbb40 (LWP 818433) "thr" 0xf7fd0579 in __kernel_vsyscall ()
    1.3  Thread 0xf73ffb40 (LWP 818434) "thr" breakpt () at thr.c:19
    2.1  Thread 0xf7dbc700 (LWP 818456) "thr" 0xf7eb2888 in clone () from /lib/libc.so.6
    2.2  Thread 0xf7dbbb40 (LWP 818457) "thr" breakpt () at thr.c:19
  * 2.3  Thread 0xf73ffb40 (LWP 818458) "thr" breakpt () at thr.c:19
  (gdb) inferior 1
  [Switching to inferior 1 [process 818430] (/home/andrew/tmp/thr)]
  [Switching to thread 1.1 (Thread 0xf7dbc700 (LWP 818430))]
  #0  0xf7eb2888 in clone () from /lib/libc.so.6
  (gdb) thread 2.2
  [Switching to thread 2.2 (Thread 0xf7dbbb40 (LWP 818457))]
  #0  breakpt () at thr.c:19
  19	  while (stop)
  (gdb)

Notice that when we switch from thread 2.3 to 1.1 using the 'inferior
1' command, GDB tells us that the inferior has changed, and that the
thread has changed (and also that the frame has changed).

But, when we switch from 1.1 to 2.2 using the 'thread 2.2' command, we
are only told about the thread change.

The 'Switching to inferior ...' line includes some useful information,
the process PID and the executable name, and I think it is a shame
that these are not presented when using the 'thread' command to switch
inferior.

So, this commit addresses this issue.

There are changes in basically two areas.  The easy part is in
thread_command (thread.c).  Here we spot when the inferior has changed
as a result of the 'thread' command, and included
USER_SELECTED_INFERIOR in the set of state passed to the
notify_user_selected_context_changed function.

The change in mi/mi-main.c is a little more involved.  In the
mi_cmd_execute function we use an instance of user_selected_context to
spot if any inferior state (frame, thread, or inferior) changes after
an MI command, this is then used to decide if there should be a call
to notify_user_selected_context_changed.

Currently, in mi_cmd_execute, notify_user_selected_context_changed is
always passed 'USER_SELECTED_THREAD | USER_SELECTED_FRAME'.  This
makes sense, the MI doesn't allow "switching inferiors" as a command,
instead, an MI frontend must switch threads, and the inferior is
switched as a consequence.  But this does mean that if a user has a
CLI and MI interpreter running, and the MI switches threads, the CLI
will only receive the thread switch style notifications, that is,
there will be no "Switching to inferior ..." line.

What I've done is rename user_selected_context::has_changed to
user_selected_context::what_changed, this function is now responsible
for returning the set of USER_SELECTED_* flags that indicate what
changed.

If anything has changed then we always return USER_SELECTED_THREAD |
USER_SELECTED_FRAME as a minimum.  This retains the existing
behaviour, but is possibly more aggressive that we need to be; the
-stack-select-frame command can only change the frame, so maybe in
this case we should only return USER_SELECTED_FRAME?  I've left that
for the future though.

However, the important change is that in ::what_changed, I now spot
when the inferior has changed and include USER_SELECTED_INFERIOR in
the set of flags that are returned.

In mi_cmd_execute we now call the new what_changed function, and use
the set of flags returned when calling
notify_user_selected_context_changed.  This means that the CLI will
now receive inferior changed notifications when appropriate.

The gdb.mi/user-selected-context-sync.exp script has been updated,
replacing the comment I quoted above with an actual test that the
inferior change is announced correctly.