gdb: fix segfault caused by using gdb-index and DWO debug information
While testing using the cc-with-gdb-index board, I discovered that GDB
would crash (segfault) on gdb.dwarf2/fission-reread.exp.

This commit is my attempt at resolving this issue, though I'm not
entirely happy with the solution I'm proposing here, it feels a little
lazy, and I'm wondering if people have thoughts on a better way to fix
it.

The Problem
===========

  1. GDB is loading information for a CU in the executable which is
     really just a stub.  The actual content has been moved into a DWO
     file.

  2. As such the per_cu->section has not been read in, instead GDB is
     reading content via dwo_unit->section.

  3. GDB ends up in dwarf2_cu::addr_type, which calls
     per_cu->addr_size (), this in turn calls
     dwarf2_per_cu::get_header (), which make the assumption that the
     per_cu->section has been read in.

  4. As the section has not been read in, the section's buffer is
     NULL, and we then try to read the CU's header from an invalid
     address.

Fixes In This Patch
===================

  1. In dwarf2_per_cu::get_header () I've added an assert that the
     section has been read in.  This changes the segfault that I was
     seeing into an assertion failure.

  2. I've extended the gdb.dwarf2/fission-reread.exp test script so
     that the test is run "normally", however the compiler builds
     things.  And then, if the compiler didn't add an index by
     default, the test is repeated adding a gdb-index, and then a
     dwarf-5 index.  This means that the bug I hit should be exposed
     without having to use the cc-with-gdb-index board.  I suspect
     most people are not testing with that board as part of their
     normal work flow.

  3. In dwarf2_cu::addr_type () I've added a call to read in the
     per_cu's section content prior to requesting the addr_size.  This
     ensures the section is read in, and mitigates the failure.

Why I Dislike This Fix
======================

Parts (1) and (2) of the patch, mentioned above, are fine.  It's
part (3) that I have a problem with.

It's worth saying that I believe this fix is functionally fine.  As I
understand it, the header of the CU in the DWO file should be
identical (at least in things like address size) to the header of the
stub CU left in the original objfile.  As a result, loading in the
stub CU, and reading the address size from there should give the
correct answer.

However, we've clearly gone to some efforts to avoid reading in the
stub CU as (I guess) a performance and memory optimisation.  I'm now
undoing that, which I dislike.

It would be better, I think, if we could avoid having to load the stub
CU, and instead could pull the required information from the DWO
objfile.

I did start trying to rework the code with that in mind, but the
change seemed to be far more complex that I had time to try and
understand right now.

The biggest thing that I didn't understand was this: I believe that
the dwarf2_cu object only exists while the DWARF is being parsed, and
is deleted after that.  While the dwarf2_per_cu object stays live
after the initial DWARF parsing, and is used for later callbacks into
the DWARF parser, e.g. for expression evaluation.

But what I don't understand is the lifetime of the DWO objects, these
seem to be attached to the dwarf2_cu, which would suggest they get
cleaned up early ... but are they also not needed later on?

What I wanted to do was move things like dwarf2_per_cu::addr_size ()
and dwarf2_per_cu::get_header () out of dwarf2_per_cu, and into
dwarf2_cu, then (at least the get_header() function) could choose to
get the header from either the per_cu, or from the dwo_unit.  But some
places dwarf2_per_cu::addr_size is called, we don't seem to have a
dwarf2_cu....

... and at that point I figured I should post this to the list and ask
for suggestions.

Conclusions
===========

The patch presented here adds an assert and improves the test script,
both of these changes are solid, and can/should be merged (IMHO).

The actual fix is, I believe, functionally correct, and could be
merged, though it might be a bit of a performance/memory hit, but
would fix an actual crash.

I'd really like peoples opinions; merge this?  Or is there a better
fix I should work towards, and what would that look like?