gdb: fixes for setting the section in minsym_found
After this commit:

  commit 6f7ad2381ae72aa592ada4a0921265aa3292b1fa
  Date:   Wed Sep 3 19:57:42 2025 +0100

      gdb: ensure bp_location::section is set correct to avoid an assert

Some issues were reported as a result of the bp_location::section
being left as NULL by the call to find_pc_overlay that was
introduced.  See this thread:

  https://inbox.sourceware.org/gdb-patches/7febb0c1-7bbd-45d5-8ebe-91c34bb4a6ce@efficios.com

The problem was that code_breakpoint::add_location relies on the
section being set in order to deduce the gdbarch.  If the section is
not set then the gdbarch is deduced using the breakpoint's gdbarch.

The bug was reported by the ROCm engineers, who have inferiors running
mixed host and GPU code, and so rely on the section being set in order
to establish the correct architecture for a specific address.

During discussion in the above thread Simon pointed out that the
change made in the above commit might not be correct anyway for
overlay debugging (does that even work, or is it used any more?), as
the commit relies on establishing a section by calling
find_pc_overlay.  However, when presented with multiple possible
sections, find_pc_overlay cannot know which section to select, and so
just picks one.  This could be different from the section of the
minimal_symbol we already had to hand.

This patch I think should (at least) resolve the issues the ROCm
engineers are seeing.

Instead of always calling find_pc_overlay I have moved the section
assignment inside the if/then/else blocks with the following
reasoning.

In the 'else' block, this is the non-function or non-ifunc case, the
address used is based on the msymbol's address, and so should be in
the same section.  In this case we can use the msymbol's section.

In the 'if' block things are more complicated.  This could be the
ifunc case, in which case func_addr could have been adjusted to a
different section, or even different objfile.

Further, when we call find_function_start_sal, we pass in just an
address, so the SAL being returned isn't going to consider which
overlay section the original msymbol was from, which could cause
problems for overlay debugging maybe?

Anyway, I'm ignoring that for now, as fixing that would be a whole big
thing.  So I'm proposing that, if find_function_start_sal returns a
symtab_and_line with a section set, then we use that section.
Otherwise, we can try to figure out a section.